Rename SelectActionGrid to HedvigGrid and make centering optional
Can be used as a generic non-lazy grid layout instead, simply by making
the item centering optional.
